requirements.txt
fastapi[standard]
pydantic_settings
asyncpg
SQLAlchemy
PyJWT
python-jose[cryptography]
bcrypt
psycopg2-binary